feat: add the renderingcontext abstraction.
it's a thin layer on top of Raylib that allows us to more easily
render either directly to the window or to a backing render texture,
without having to call Raylib functions directly, or manage the lifetimes
of its objects.

this is in preparation for a new editor UI, where the scene is just
a window inside of the editor.