feat+refactor!: use renderingcontexts, overhaul editor UI.
the editor can now take over the entire rendering of the window,
allowing us to draw the editor outside of the stage itself.

this commit also makes the stage itself get drawn as an imgui window
inside of the editor.

furthermore, the editor is now a resizable window, with added docking.
(the default layout is not yet finished)