feat: start working on the asset database system!
this will be the main store for all of the assets, so we can reuse them
efficiently, without having to reload them every single time we instantiate
a component with an asset that was previously used.

this also adds the ability to specify properties on assets and makes them
more abstract, not relying on other components to modify them.