feat+refactor!!!: turn the texture class into an asset (and more).
this is a MASSIVE refactor of how we handle textures in VTubeTiny
instead of every element managing its own textures, we now delegate that
to the asset database system.

instead of dropping textures onto component gui elements, we now have
a dropdown where you can pick from the currently loaded in assets.

the asset browser window now shows the currently loaded assets, alongside
with their previews and internal configuration.

in order to load in new assets, you can now drop the file onto the asset
browser, making it instantly load in the background.

this also renders VTubeTiny not compatible with earlier configuration files.