feat+refactor!!: add Spout broadcasting support.
i am not 100% satisfied with this, but for now it can work.

you can now toggle the spout broadcasting option, and the stage will
be broadcasted both from the editor and from the proper stage viewer.

this preserves transparency and allows for using the output of vtubetiny
in many applications, without having to resort to window capture and
chroma keying.

this is only a x64 windows thing, so on other platforms enabling spout
will just do nothing.

i don't like the fact that i had to add a fullscreen blitting thing to
the rendering context abstraction, but we need to draw the stage to the
screen when we're in the stage viewer. maybe i'll revisit it later.