refactor: launch into the editor by default.
this change is mostly motivated by the fact, that the default
before when launched with no parameters was to launch in the stage
viewer mode with an empty stage.

this was a leftover from when there was no editor mode, and you had
to create a stage yourself via a text editor. but now that we can
create stages right within VTubeTiny, launching into an empty window
without an ability to do anything is a bit counter-intuitive to new
users.

so instead, the default solution will be to launch to the editor
itself, moving the stage viewer to the "-s" parameter.