feat+refactor!!: turn the character into an asset.
this is a HUGE change, but bear with me.

the main reasoning for it, is that i don't believe character configs
(which is the entire texture setup) should be tied to the animator component.
removing an asset should not remove the character. they're first class citizens
in a vtuber software, after all.

so, instead, i've split the AnimatorCharacter into two independent parts:

1. the Character asset.
   this one contains all of the visual definitions for the character,
   all the textures for the different states and etc.

2. the CharacterAnimator class.
   this one deals with using the information from the Character and
   animating the texture states.

the new arrangement feels much more sane in my opinion.